requirements.txt
anthropic>=0.25.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
gitpython>=3.1.0
jsonschema>=4.0.0
colorlog>=6.7.0
websockets>=12.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
